+++ zcash_primitives/src/transaction/fees/fixed.rs
@@ -1,7 +1,7 @@
 use crate::{
     consensus::{self, BlockHeight},
     transaction::components::amount::NonNegativeAmount,
-    transaction::fees::{transparent, zip317},
+    transaction::fees::transparent,
 };
 
 #[cfg(zcash_unstable = "zfuture")]
@@ -16,27 +16,12 @@ pub struct FeeRule {
 
 impl FeeRule {
     /// Creates a new nonstandard fixed fee rule with the specified fixed fee.
+    #[deprecated(
+        note = "Using a fixed fee may result in a transaction that cannot be mined. \
+                To calculate the ZIP 317 fee, use `transaction::fees::zip317::FeeRule::standard()`."
+    )]
     pub fn non_standard(fixed_fee: NonNegativeAmount) -> Self {
         Self { fixed_fee }
-    }
-
-    /// Creates a new fixed fee rule with the minimum possible [ZIP 317] fee,
-    /// i.e. 10000 zatoshis.
-    ///
-    /// Note that using a fixed fee is not compliant with [ZIP 317]; consider
-    /// using [`zcash_primitives::transaction::fees::zip317::FeeRule::standard()`]
-    /// instead.
-    ///
-    /// [`zcash_primitives::transaction::fees::zip317::FeeRule::standard()`]: crate::transaction::fees::zip317::FeeRule::standard
-    /// [ZIP 317]: https://zips.z.cash/zip-0317
-    #[deprecated(
-        since = "0.12.0",
-        note = "To calculate the ZIP 317 fee, use `transaction::fees::zip317::FeeRule::standard()`. For a fixed fee, use the `non_standard` constructor."
-    )]
-    pub fn standard() -> Self {
-        Self {
-            fixed_fee: zip317::MINIMUM_FEE,
-        }
     }
 
     /// Returns the fixed fee amount which this rule was configured.